A few updates
* Added sweet new Menu Filter to get right to the darn things
* Swapped out how Nav & Versions were being generated to create JSX for React to consume
* Updated gitignore to make sure generated data for site is not included in repo `*`

`*` Trying hard to make sure anything that had a copyright in it is not being stored in this repository to limit who can just download everything.  Current setup is that you will need access to the ZIP files, which are loosely behind an HTTP basic auth, and while it's fine if you figured out how to get them ( they were accessible online at one time, after all ), most people won't be bothered, as the only other way to get the content, is to clone ths repo and build the entire project, then spent time hacking at it to stop it from trashing the stuff you just downloaded, ha ha.