After testing this out a bit, gonna have to limit this to a single version :(
However, I can make multiple sites and set them up to know about each other. I am just not happy with how large the generated site is with all the content, especially since most users won't use more than one.

I was hoping I could shard each version into smaller pieces and bring them in as needed, but the way this app works, it needs to compile all docs into an index if it's in URL path.  So that kind of broke that option, and keeping meant the first initial page load was like 4MB or something like that, which is just as bad aas the current docs.

This solution keeps it around 200KB, which is super acceptable :)