Pushing up changes from long weekend
Might still try to wrap up some more later tonight.  But this update has a few things

* Switching out how versions are loaded since we can't load them all on the same site
* Simplifying how version info is fetched
* Fixing CLI scripts to support version structure change ( DIFFs still need ALL versions to make DIFF, but website will only show one version ( switcher will be updated to just take you to these other versions vs trying to load them all on the same site )
* Making a new external version JSON file that this loads so each installation can know which ZIP files are on the central server

NOTE: These changes are being made in order to keep this stuff as light weight as possible.  Once a version of a site is rendered, it should be static anyway.  Regarding DIFF's I think I will make this so it can still compare with future versions, and just link to each version in the timeline.